agents/01_pm_support_bot/requirements.txt
strands-agents
bedrock-agentcore
boto3
python-dotenv